Avoid heap-buffer-overflow if xdata of patch object is a row vector (bug #66466).
* libinterp/corefcn/graphics.cc (patch::properties::update_fvc): The "xdata"
property of a patch graphics object is allowed to be a row vector (in which
case it is treated like a column vector). In this case, check whether colors
are set per vertex after the input has been transformed to column vectors. This
avoids a potential heap-buffer-overflow when creating the "facevertexcdata" of
the patch graphics object.